+++ CHANGELOG.md
@@ -5,15 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+---
 ## [0.0.4] - Unreleased
+
+### Added
+- Line extension: append any String after a path (comment, file details, etc.)
 
 ### Changed
 - Filtering: moved to options
+- Compact dirs: root dir is now never compacted
 
 ### Removed
 - Error handling: exceptions are now thrown instead of being handled by the renderer
 
-
+---
 ## [0.0.3] - 2025-09-21
 
 ### Added
@@ -25,14 +30,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sorting: option method `withFileSort` renamed in `sort`
 - Child limit: various renaming and refactor
 
-
+---
 ## [0.0.2] - 2025-09-16
 
 ### Added
 - Option: sorting files and directories
 - 39 new default files & extension mappings for emojis
 
-
+---
 ## [0.0.1] - 2025-09-14
 
 ### Added